requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.109.0
filelock==3.18.0
fsspec==2025.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.39.0
google-genai==1.11.0
google-generativeai==0.3.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.63.0rc1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.25.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.2
pydantic_core==2.14.5
PyPDF2==3.0.1
python-docx==1.0.0
python-dotenv==1.0.0
python-multipart==0.0.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sniffio==1.3.1
starlette==0.35.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.0
websockets==15.0.1